Split files, fix nullability warnings
